requirements.txt
Python 3.10.0
Django==4.1.3
psycopg2==2.9.5
PyJWT==2.6.0